* changes:
  android: Fix misc compiler errors for GCC 5.4+
  build-breakpad.sh: Fix script
  build: Disable link-time optimization, add --lto option.

+++ android/android-emu/android/base/containers/PointerSet_unittest.cpp
@@ -192,7 +192,7 @@ TEST(PointerSet, Iterator) {
         EXPECT_TRUE(iter.hasNext());
         Foo* foo = iter.next();
         EXPECT_TRUE(foo);
-        ASSERT_LT(foo->value(), kCount);
+        ASSERT_LT(foo->value(), (int)kCount);
         EXPECT_FALSE(flags[foo->value()]) << "Item #" << foo->value();
         flags[foo->value()] = true;
     }
