+++ android/android-emu/android/emulation/AdbDebugPipe_unittest.cpp
@@ -29,7 +29,8 @@ TEST(AdbDebugPipe, noOutput) {
     auto guest = TestAndroidPipeDevice::Guest::create();
     EXPECT_EQ(0, guest->connect("qemud:adb-debug"));
     constexpr StringView kMessage = "Hello debug world!";
-    EXPECT_EQ(kMessage.size(), guest->write(kMessage.c_str(), kMessage.size()));
+    EXPECT_EQ(static_cast<ssize_t>(kMessage.size()),
+              guest->write(kMessage.c_str(), kMessage.size()));
 }
 
 TEST(AdbDebugPipe, withOutput) {
@@ -40,7 +41,8 @@ TEST(AdbDebugPipe, withOutput) {
     auto guest = TestAndroidPipeDevice::Guest::create();
     EXPECT_EQ(0, guest->connect("qemud:adb-debug"));
     constexpr StringView kMessage = "Hello debug world!";
-    EXPECT_EQ(kMessage.size(), guest->write(kMessage.c_str(), kMessage.size()));
+    EXPECT_EQ(static_cast<ssize_t>(kMessage.size()),
+              guest->write(kMessage.c_str(), kMessage.size()));
 
     EXPECT_EQ(kMessage, outStream->view());
 }
